+++ src/components/TestTips/TestTips.js
@@ -27,8 +27,8 @@ class TestTips extends Component {
     handleSubmit(e) {
         e.preventDefault();
 
-        console.log('The form was submitted with the following data:');
-        console.log(this.state);
+        //console.log('The form was submitted with the following data:');
+        //console.log(this.state);
     }
 
     render() {
